--- app/(auth)/_layout.tsx
@@ -1,12 +1,0 @@
-// import { Redirect, Stack } from 'expo-router';
-// import { useAuth } from '@clerk/clerk-expo';
-
-// export default function AuthRoutesLayout() {
-//   const { isSignedIn } = useAuth();
-
-//   if (isSignedIn) {
-//     return <Redirect href={'/'} />;
-//   }
-
-//   return <Stack />;
-// }
