+++ src/components/buttons/CustomButton.tsx
@@ -1,29 +1,25 @@
-import { Pressable, PressableProps, Text } from 'react-native';
+import { PropsWithChildren } from 'react';
+import { Pressable, PressableProps, StyleProp, ViewStyle } from 'react-native';
 import { StyleSheet } from 'react-native-unistyles';
 
 type CustomButtonProps = {
-  // custom props
-  text?: string;
+  style?: StyleProp<ViewStyle>;
 } & PressableProps;
 
-export const CustomButton = ({ text, style, ...pressableProps }: CustomButtonProps) => {
+export const CustomButton = ({
+  children,
+  style,
+  ...pressableProps
+}: PropsWithChildren<CustomButtonProps>) => {
   return (
-    <Pressable {...pressableProps} style={[styles.signInButton]}>
-      {text && <Text style={[styles.signInButtonText]}>{text}</Text>}
+    <Pressable {...pressableProps} style={[styles.button, style]}>
+      {children}
     </Pressable>
   );
 };
 
 const styles = StyleSheet.create(theme => ({
-  signInButton: {
-    backgroundColor: theme.colors.elevatedSurface,
-    padding: 15,
-    borderWidth: 1,
-    borderRadius: 5,
-    alignItems: 'center',
-  },
-  signInButtonText: {
-    color: theme.colors.bodyText,
-    fontSize: theme.fontSizes.medium,
+  button: {
+    display: 'flex',
   },
 }));
